BackEnd/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-doc==0.0.2
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.14.2
bleach==6.3.0
build==1.3.0
cachetools==6.2.1
certifi==2023.7.22
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.2.1
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.3
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.8.0
docopt==0.6.2
durationpy==0.10
ecdsa==0.19.1
edge-tts==6.1.9
email-validator==2.3.0
executing==2.2.1
fastapi==0.120.0
fastjsonschema==2.21.2
filelock==3.20.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.9.0
git-filter-repo==2.47.0
google-auth==2.41.1
google-genai==1.46.0
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
kubernetes==34.1.0
langchain==1.0.2
langchain-classic==1.0.0
langchain-community==0.4
langchain-core==1.0.0
langchain-huggingface==1.0.0
langchain-text-splitters==1.0.0
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.38
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.1.4
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.5
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.23.2
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.3
ormsgpack==1.11.0
overrides==7.7.0
packaging==25.0
pandocfilters==1.5.1
parso==0.8.5
passlib==1.7.4
pickleshare==0.7.5
pillow==12.0.0
pipreqs==0.5.0
platformdirs==4.5.0
posthog==5.4.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.5
pyotp==2.9.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
pyzmq==27.1.0
qrcode==8.2
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
resend==2.17.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.48.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.1
torch==2.9.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.3.0
user-agents==2.2.0
uvicorn==0.38.0
watchfiles==1.1.1
wcwidth==0.2.14
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
xxhash==3.6.0
yarg==0.1.9
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
